image-resizer/requirements.txt
flask==2.3.3
Pillow==10.0.0
gunicorn==21.2.0